+++ tdas/heap.h
@@ -5,7 +5,7 @@ typedef struct Heap Heap;
 
 void* heap_top(Heap* pq);
 
-void heap_push(Heap* pq, void* data, int priority);
+void heap_push(Heap* pq, void* data, double priority);
 
 void heap_pop(Heap* pq);
